:bug: fix duplicate keys
Card components, Cart Item components, and the list of ingredients were
using the same keys. Sometimes duplicate cards would render because
of this. I figured for the scope of this project, just adding some
numbers to the keys should be enough to fix the issue and that none
of them should clash.