:bug: actually fix the duplicate key error
Okay now THIS time I think I actually found the culprit. I'd decided to
snoop around and look through the Redux Dev Tools to see if I could
catch the sneaky little error in the act. And I did. Turns out it was
the state itself duplicating the data inside, not necessarily React
accidentally rerendering components.

What I was initially doing was fetching the data, and pushing the
destructured payload into the state. BUT, every time it refetched the data,
it would just push the same data in! But now I actually set the
state to an array of the destructured payload, so that it can overwrite it
every single time data is refetched. Phew.