:iphone: make cart wider on mobile
I'd made the cart pretty narrow on smaller screen widths since I'm
developing the site on a desktop browser and the scrollbar is pretty
wide. But I forgot that on mobile devices/browsers, the scrollbar is
no where near as large. When accessiing the site on my phone, it was
a lot more apparent!